requirements.txt
python-telegram-bot==21.7
ankr-sdk==1.0.5
pandas==2.2.3
apscheduler==3.10.4
gspread==6.1.4
google-auth==2.37.0
python-dotenv==1.0.1
matplotlib==3.9.2
requests==2.32.3
